Use a long for the content-length attribute

+++ ng-appserver/src/main/java/ng/appserver/NGResourceRequestHandlerDynamic.java
@@ -68,5 +68,5 @@ public class NGResourceRequestHandlerDynamic extends NGRequestHandler {
 	public static record NGDynamicResource(
 			InputStream inputStream,
 			String mimeType,
-			Integer length ) {}
+			Long length ) {}
 }
